# Calculate request stats directly rather than asyncronous.